Refactor ArticlePage to direct users to error page when trying to have direct access to an article page
This commit fixes the issue of an API error occurring when a user tries to access an article through direct access. Because the implementation uses useLocation, a blog aricle is only accessible by navigating through an article card from the blog page.Now, if a user attempts to access an article directly, they will be redirected to the error page instead of being faced with an API error message.